refactor: move MocketEntry from mocket.mocket to mocket.entry
